area-processor/requirements.txt
PyMuPDF>=1.23.0
opencv-python-headless>=4.8.0
shapely>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
supabase>=2.0.0
python-dotenv>=1.0.0
